--- networkdispatchers/src/main/AndroidManifest.xml
@@ -1,4 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.ninchat.sdk">
-    <uses-permission android:name="android.permission.INTERNET" />
-</manifest>
