--- networkdispatchers/src/main/res/values/strings.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="ninchat_site_secret">dummy_value</string>
-    <string name="ninchat_site_secret_wrong">dummy_value</string>
-    <string name="ninchat_server_address">dummy_value</string>
-    <string name="ninchat_server_address_wrong">dummy_value</string>
-    <string name="ninchat_configuration_key">dummy_value</string>
-    <string name="ninchat_configuration_key_wrong">dummy_value</string>
-    <string name="ninchat_queue">dummy_value</string>
-    <string name="ninchat_queue_wrong">dummy_value</string>
-</resources>
