+++ ninchatsdk/src/androidTest/java/com/ninchat/sdk/networkdispatchers/NinchatOpenSessionTest.kt
@@ -1,10 +1,11 @@
-package com.ninchat.sdk
+package com.ninchat.sdk.networkdispatchers
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry.getInstrumentation
 import com.ninchat.client.Props
 import com.ninchat.client.Session
 import com.ninchat.client.SessionEventHandler
+import com.ninchat.sdk.R
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
