+++ ninchatsdk/src/main/java/com/ninchat/sdk/networkdispatchers/NinchatDescribeRealmQueues.kt
@@ -1,19 +1,25 @@
-package com.ninchat.sdk
+package com.ninchat.sdk.networkdispatchers
 
 import com.ninchat.client.Props
 import com.ninchat.client.Session
+import com.ninchat.client.Strings
 import kotlinx.coroutines.channels.SendChannel
 
-class NinchatDescribeFile {
+class NinchatDescribeRealmQueues {
     companion object {
         suspend fun execute(currentSession: Session? = null,
                             channel: SendChannel<Long>,
-                            fileId: String? = null) {
+                            realmId: String? = null,
+                            audienceQueues: Collection<String>) {
+
             val params = Props()
-            params.setString("action", "describe_file");
-            fileId?.let {
-                params.setString("file_id", fileId);
+            params.setString("action", "describe_realm_queues")
+            params.setString("realm_id", realmId)
+            val queues = Strings()
+            for (queue in audienceQueues) {
+                queues.append(queue)
             }
+            params.setStringArray("queue_ids", queues)
             val actionId: Long = try {
                 currentSession?.send(params, null) ?: -1
             } catch (e: Exception) {
