+++ settings.gradle
@@ -1,2 +1,1 @@
-include ':networkdispatchers'
 include ':ninchatsdk'
